+++ lib/Controller/BookmarkController.php
@@ -146,9 +146,21 @@ class BookmarkController extends ApiController {
 	 */
 	private function _returnBookmarkAsArray(Bookmark $bookmark): array {
 		$array = $bookmark->toArray();
+		$array['folders'] = array_values(array_filter($this->treeMapper->findParentsOf(TreeMapper::TYPE_BOOKMARK, $bookmark->getId()), function (Folder $folder) {
+			if ($folder->getUserId() === $this->authorizer->getUserId()) {
+				return true;
+			}
+			if ($this->toExternalFolderId($folder->getId()) === -1) {
+				return false;
+			}
+			if ($this->authorizer->getUserId() !== null && $this->folders->findShareByDescendantAndUser($folder, $this->authorizer->getUserId()) !== null) {
+				return true;
+			}
+			return false;
+		}));
 		$array['folders'] = array_map(function (Folder $folder) {
 			return $this->toExternalFolderId($folder->getId());
-		}, $this->treeMapper->findParentsOf(TreeMapper::TYPE_BOOKMARK, $bookmark->getId()));
+		}, $array['folders']);
 		$array['tags'] = $this->tagMapper->findByBookmark($bookmark->getId());
 		return $array;
 	}
@@ -462,7 +474,7 @@ class BookmarkController extends ApiController {
 					return new JSONResponse(['status' => 'error', 'data' => 'Insufficient permissions'], Http::STATUS_BAD_REQUEST);
 				}
 			}
-			$bookmark = $this->bookmarks->update($id, $url, $title, $description, $tags, $folders);
+			$bookmark = $this->bookmarks->update($this->authorizer->getUserId(), $id, $url, $title, $description, $tags, $folders);
 			return new JSONResponse(['item' => $bookmark ? $this->_returnBookmarkAsArray($bookmark) : null, 'status' => 'success']);
 		} catch (AlreadyExistsError $e) {
 			// This is really unlikely, as we make sure to use the existing one if it already exists