+++ src/components/Breadcrumbs.vue
@@ -199,24 +199,14 @@ export default {
 </script>
 <style>
 .breadcrumbs {
-	padding: 2px 8px;
+	padding: 0 8px 0 44px;
 	display: flex;
-	position: fixed;
+	position: absolute;
 	z-index: 100;
 	background: var(--color-main-background-translucent);
-	right: 0;
-	left: 300px;
-}
-
-@media only screen and (max-width: 1024px) {
-	.breadcrumbs {
-		padding-left: 52px;
-		left: 0;
-	}
-}
-.breadcrumbs.wide {
-	padding: 2px 8px;
 	left: 0;
+	right: 0;
+	top: 0;
 }
 
 .breadcrumbs + * {
@@ -232,7 +222,7 @@ export default {
 .breadcrumbs__path > * {
 	display: inline-block;
 	height: 30px;
-	padding: 7px;
+	padding: 5px 7px;
 	flex-shrink: 0;
 }